fix remove file, get template menus

+++ app/order-api/src/file/file.service.ts
@@ -36,7 +36,7 @@ export class FileService {
   async uploadFiles(files: Express.Multer.File[]) {
     const context = `${FileService.name}.${this.uploadFiles.name}`;
     const uploadedFiles = await Promise.all(
-      files.map((file) => this.saveFile(file))
+      files.map((file) => this.saveFile(file)),
     );
     this.logger.log(`Files uploaded successfully`, context);
     return uploadedFiles;
@@ -60,38 +60,43 @@ export class FileService {
 
   public async removeFile(filename?: string): Promise<void> {
     const context = `${FileService.name}.${this.removeFile.name}`;
-    const file = await this.fileRepository.findOne({
-      where: { name: filename },
-    });
-    if (!file) return;
-    await this.fileRepository.remove(file);
+    if (filename) {
+      const file = await this.fileRepository.findOne({
+        where: { name: filename },
+      });
+      if (!file) return;
+      await this.fileRepository.remove(file);
+    }
     this.logger.log(`File ${filename} removed successfully`, context);
   }
 
-  public handleDuplicateFilesName(files: Express.Multer.File[]): Express.Multer.File[] {
+  public handleDuplicateFilesName(
+    files: Express.Multer.File[],
+  ): Express.Multer.File[] {
     const fileNameCount: { [key: string]: number } = {};
     const renamedFiles: Express.Multer.File[] = [];
-  
+
     files.forEach((file) => {
       const fileExtension = file.originalname.split('.').pop();
-      const baseName = file.originalname.replace(/\.[^/.]+$/, ''); 
-  
+      const baseName = file.originalname.replace(/\.[^/.]+$/, '');
+
       if (fileNameCount[baseName]) {
         fileNameCount[baseName]++;
       } else {
         fileNameCount[baseName] = 1;
       }
-  
-      const newName = fileNameCount[baseName] === 1
-        ? file.originalname
-        : `${baseName}(${fileNameCount[baseName] - 1}).${fileExtension}`;
-  
+
+      const newName =
+        fileNameCount[baseName] === 1
+          ? file.originalname
+          : `${baseName}(${fileNameCount[baseName] - 1}).${fileExtension}`;
+
       renamedFiles.push({
         ...file,
         originalname: newName,
       });
     });
-  
+
     return renamedFiles;
   }
 }
